Replace insertSpacing() with addSpacing()
insertSpacing() will cause Flame to crash with no useful error message
if the index value is not valid.

Change to addSpacing() which renders the same and is easier to read.

It's easier to read because the addSpacing() is placed where you expect
the spacing to be added.  insertSpacing() can be placed anywhere inside
a QBoxLayout and the index refers to the order in which addLayout is
used.  So, insertSpacing() does not have to be where you expect the
spacing to render.  Unsure why insertSpacing() was originally chosen.